Setting up to include audio events
