+++ Powerup/Sounds.xcassets/silence.dataset/Contents.json
@@ -6,7 +6,7 @@
   "data" : [
     {
       "idiom" : "universal",
-      "filename" : "placeholder2.mp3"
+      "filename" : "silence.mp3"
     }
   ]
 }
